Migrate READY handling to a separate file
In preparation for moving all high-level gateway event handling to a
separate location, move the code handling the READY event into a
separate file.

Ultimately, I'm not too sure that converting class methods into exported
free functions is the best idea, especially since they take an instance
of the class and mutate it to boot---but the class itself was getting
pretty long and it made the file a bit difficult to navigate.