use PersistentWS for gateway connections (#16)
Use the new PersistentWS abstraction from the platform SDK for our
connection to the gateway. Hopefully this uncomplicates our reconnection
logic somewhat.

* bump platform-sdk

* use PersistentWS

* use Array.prototype.at

Co-authored-by: Kishan Bagaria <1093313+KishanBagaria@users.noreply.github.com>

* remove superfluous Promise.resolve

Co-authored-by: Kishan Bagaria <1093313+KishanBagaria@users.noreply.github.com>

* only create PersistentWS once

* pass null to oldestCursor

If `hasMore` is `false`, then `oldestCursor` can be `null`/`undefined`.

* log spurious connect calls to sentry

* don't error when sending before ready

PersistentWS already queues sends for us, so we don't need to handle
this ourselves.

* remove WSClient._ready

This property is redundant as its value matches the value of
PersistentWS.connected most of the time. We can simply use that.

* random cosmetic change

* nuke unneeded connected check

Co-authored-by: Kishan Bagaria <1093313+KishanBagaria@users.noreply.github.com>

---------

Co-authored-by: Kishan Bagaria <1093313+KishanBagaria@users.noreply.github.com>