provide strong typing for gateway event handling
When ingesting incoming Gateway events via event emitter, have the types
correspond to the actual (believed) shape of the data. This way, the app
code can be completely free of casts, and we can encode the data shape
invariants elsewhere.

This is done by subclassing EventEmitter, and defining a new "lookup
table" type that maps each message type to the type of event data it
receives. Gateway messages that are emitted then take on the
corresponding type automatically.

Changes were made to the mapping code so that all of the types line up.
Further verification of the event data types can come in a future patch,
as I'm still not 100% confident that they align with reality.