completely reimplement resumption and reconnection & remove polling (#18)
* bump platform-sdk

* remove polling and revamp reconnection logic

Remove behavior where we poll the HTTP API when we think we're offline.
This is incompatible with `PersistentWS`'s connection lifecycle
behavior, and can cause an explosion of reconnection spam that
Cloudflare-bans the user.

Implement proper RESUME support. I'm not sure if we were doing this
correctly before. Overall, `WSClient` has become much more behaved. I
did some preliminary smoke testing and it responds correctly to having
Wi-Fi shut off and the machine going to sleep.

After landing patches that add further customizability to how
`PersistentWS` behaves, we now rely on it to perform low-level WebSocket
reconnections whenever appropriate. We track the necessary state to tell
it whether to reconnect or not, whether we should RESUME or IDENTIFY,
and which gateway URL to use (when resuming, it's different).

To help facilitate debugging, this code introduces a `texts.log` wrapper
that also outputs a timestamp, which is useful when debugging. This
should really be integrated into PAS/desktop instead, though.

* ignore reconnectRealtime

Forcibly disconnecting the WebSocket while automatic reconnection logic
is trying to take care of things can result in errors.

* close with a permitted code when zombifying

The `ws` library doesn't let us use 1006, so use a code that aligns with
Discord's developer guidance and doesn't throw. We can investigate which
code Discord's clients actually use later on.